% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCommas.R
\name{addCommas}
\alias{addCommas}
\title{Add commas to a large number}
\usage{
addCommas(numbers)
}
\arguments{
\item{numbers}{Vector of non-negative numbers (will be rounded to integers)}
}
\value{
Character string with numbers written like \code{"7,547,085"}.
}
\description{
Convert a number to a string, with commas every 3rd digit
Convert a number to a string, with commas every 3rd digit
}
\examples{
\dontrun{
addCommas(c(231, 91310, 2123, 9911001020, 999723285))
}
}
