% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMean.R
\name{colMean}
\alias{colMean}
\title{colMean}
\usage{
colMean(df_col, removeNA = TRUE)
}
\arguments{
\item{df_col}{A dataframe column.}

\item{removeNA}{Logical. Sould NAs be removed? The defualt is TRUE.}
}
\description{
Find the mean of a numeric column.
}
\examples{

\dontrun{
mean <- colMean(df$col)
}
}
