% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exit}
\alias{exit}
\title{exit R without saving
exit R without saving workspace.}
\usage{
exit()
}
\description{
exit R without saving
exit R without saving workspace.
}
\details{
This just calls \code{q("no")}
}
