% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFile.R
\name{openFile}
\alias{openFile}
\title{Open a file}
\usage{
openFile(file)
}
\arguments{
\item{file}{File name (character string)}
}
\description{
Open a file using \code{\link[base]{system}} and \code{open}.
}
\examples{
\dontrun{openFile("myplot.pdf")}
}
