% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMoney.R
\name{printMoney}
\alias{printMoney}
\title{Print numeric as a dollar figure}
\usage{
printMoney(value, currency.sym = "$", digits = 2, sep = ",",
  decimal = ".")
}
\arguments{
\item{value}{Vector of non-negative numbers (will be rounded to integers)}

\item{currency.sym}{The BigMark you would like to use.}

\item{digits}{How many digits after the point sep.}

\item{sep}{The thousands seperator you would like to use.}

\item{decimal}{The mark you would like to use for the decimal point.}
}
\value{
Character string with numbers written like \code{"$7,547,085"}.
}
\description{
Print money with dollar signs and commas
}
\examples{
\dontrun{
printMoney(999723285)
}
}
