% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers}
\usage{
removeOutliers(x, na.rm = TRUE)
}
\arguments{
\item{x}{A data frame}

\item{na.rm}{Logical: remove NA.}
}
\description{
Remove outliers for the purpose of plotting
}
\examples{
\dontrun{
foo <- removeOutliers(bar)
}
}
