% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimWhite.R
\name{trimWhite}
\alias{trimWhite}
\title{trimWhite}
\usage{
trimWhite(x)
}
\arguments{
\item{x}{A dataframe, matrix or vector.}
}
\description{
Trim leading or trailing whitespace from.
}
\examples{

\dontrun{
trimed_df <- trimWhite(df)
}
}
