% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitPost.R
\name{KnitPost}
\alias{KnitPost}
\title{Knit Rmd to Md}
\usage{
KnitPost(fileName = NULL, category = NULL)
}
\arguments{
\item{fileName}{The Rmd file you would like to knit.}

\item{category}{The blog category where the post should be placed.}
}
\description{
Knit an Rmd file to an Md that is suitable for a Jekyll blog.
}
\examples{

\dontrun{
KnitPost("2017-05-18-test.Rmd")
}
}
